bhoonidhi_backend/requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=1.0.0

# Flask and extensions
Flask>=2.3.0
Flask-CORS>=4.0.0

# Geospatial and satellite data processing
rasterio>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
pyproj>=3.4.0

# AI/ML for image analysis
transformers>=4.30.0
torch>=2.0.0

# LangChain for agentic capabilities
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# Streamlit for frontend
streamlit>=1.28.0

# Optional: for production deployment
gunicorn>=21.0.0

